Create File Reader (ASCII Art optional))
This is a program that is supposed to read the characters of a file and print those characters from their initial integer values (in bytes) from the read method. 

I personally used ASCII Art in my "inspect.txt" file to print Inspector gadget. Follow along and copy and paste some ASCII Art of your choice in a txt file if you would like!

Code layout from "Bro Code" on Youtube.